Avoid spurious comment highlighting
The previous comment highlighting incorrectly highlighted the line
following `#` characters occurring within a cell/token as comments, in
cases such as `foo#bar` and `foo #bar`.

The `#bar` strings above should not be highlighted as comments, as they
do not describe the start of a cell or token, which are delimited by at
least either two spaces or a tab.

<http://robotframework.org/robotframework/latest/RobotFrameworkUserGuide.html#rules-for-parsing-the-data>